Improve single-`char` unmappable search
