+++ test/jdk/sun/nio/cs/TestEncoderReplaceUTF16.java
@@ -27,6 +27,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import sun.nio.cs.ArrayEncoder;
 
+import java.nio.ByteBuffer;
+import java.nio.CharBuffer;
 import java.nio.charset.Charset;
 import java.nio.charset.CharsetEncoder;
 import java.nio.charset.CoderResult;
@@ -36,6 +38,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.Predicate;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /*
@@ -146,18 +149,17 @@ class TestEncoderReplaceUTF16 {
         }
 
         // Try to find a single-`char` unmappable
-        Predicate<char[]> isUnmappable1 = TestEncoderReplaceLatin1.createUnmappableSingleCharPredicate(encoder);
         char[] unmappable1 = {0};
-        for (int i = 0xFF; i < 0xFFFF; i++) {
-            unmappable1[0] = (char) i;
-            boolean unmappable = isUnmappable1.test(unmappable1);
+        for (char c = 0xFF; c < 0xFFFF; c++) {
+            unmappable1[0] = c;
+            boolean unmappable = !encoder.canEncode(c);
             if (unmappable) {
                 return unmappable1;
             }
         }
 
         // Try to find a double-`char` unmappable
-        Predicate<char[]> isUnmappable2 = TestEncoderReplaceLatin1.createUnmappableDoubleCharPredicate(encoder);
+        Predicate<char[]> isUnmappable2 = createUnmappableDoubleCharPredicate(encoder);
         char[] unmappable2 = {0, 0};
         for (int i = 0xFF; i < 0xFFFF; i++) {
             unmappable2[0] = (char) i;
@@ -172,6 +174,18 @@ class TestEncoderReplaceUTF16 {
 
         System.err.println("Could not find an unmappable character!");
         return null;
+    }
+
+    static Predicate<char[]> createUnmappableDoubleCharPredicate(CharsetEncoder encoder) {
+        assertEquals(encoder.unmappableCharacterAction(), CodingErrorAction.REPORT);
+        CharBuffer charBuffer = CharBuffer.allocate(2);
+        ByteBuffer byteBuffer = ByteBuffer.allocate(4);
+        return unmappable -> {
+            charBuffer.clear().put(unmappable).flip();
+            byteBuffer.clear();
+            CoderResult coderResult = encoder.encode(charBuffer, byteBuffer, true);
+            return coderResult.isUnmappable();
+        };
     }
 
     private static byte[] utf16Bytes(char[] cs) {
